Add OnFileRemoving life cycle hook #106

+++ src/Serilog.Sinks.File/Sinks/File/RollingFileSink.cs
@@ -199,11 +199,12 @@ namespace Serilog.Sinks.File
                 var fullPath = Path.Combine(_roller.LogFileDirectory, obsolete);
                 try
                 {
-                    System.IO.File.Delete(fullPath);
+                    if (_hooks == null || _hooks.OnFileRemoving(fullPath))
+                        System.IO.File.Delete(fullPath);
                 }
                 catch (Exception ex)
                 {
-                    SelfLog.WriteLine("Error {0} while removing obsolete log file {1}", ex, fullPath);
+                    SelfLog.WriteLine("Error {0} while processing obsolete log file {1}", ex, fullPath);
                 }
             }
         }
